Add (Front-end): popup confirm to delete event/contributor

+++ client/src/app/components/contributor/contributor.component.ts
@@ -4,6 +4,8 @@ import { ContributorService } from '../../services/contributor.service';
 import { ContributorClass } from '../../domain/contributor.class';
 import { RepresentedOnMapComponent } from '../base/represented-on-map/represented-on-map.component';
 import { AuthenticationService } from '../../services/authentication.service';
+import { DialogComponent, DialogParams } from '../common/dialog/dialog.component';
+import { MatDialog } from '@angular/material';
 
 @Component({
   selector: 'app-contributor',
@@ -14,11 +16,11 @@ export class ContributorComponent extends RepresentedOnMapComponent implements O
 
   contributor: ContributorClass;
 
-  constructor(
+  constructor(private dialog: MatDialog,
     private route: ActivatedRoute,
     private router: Router,
-    private contributorService: ContributorService,
-    public authService: AuthenticationService
+    public authService: AuthenticationService,
+    private contributorService: ContributorService
   ) {
     super();
   }
@@ -31,14 +33,25 @@ export class ContributorComponent extends RepresentedOnMapComponent implements O
     });
   }
 
-  deleteContributor(contributorId) {
-    this.contributorService.delete(contributorId).subscribe(res => {
-      this.removePoint.emit({type: ContributorClass, id: contributorId});
-      this.router.navigate(['/admin']);
-    })
+  public deleteContributor(contributorId) {
+    this.dialog.open<DialogComponent, DialogParams>(DialogComponent, {
+      width: '550px',
+      panelClass: 'dialog',
+      data: {
+        title: 'Suppression',
+        body: 'Êtes-vous sûr de vouloir supprimer ce contributeur ?'
+      }
+    }).afterClosed().subscribe((res: boolean) => {
+      if (res) {
+        this.contributorService.delete(contributorId).subscribe(res => {
+          this.removePoint.emit({ type: ContributorClass, id: contributorId });
+          this.router.navigate(['/admin']);
+        });
+      }
+    });
   }
 
-  modifyContributor(contributorId) {
+  public modifyContributor(contributorId) {
     this.router.navigate(['admin', 'contributors', contributorId, 'edit']);
   }
 }
